requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Financial Data Sources
yfinance>=0.2.0
pandas-datareader>=0.10.0
alpha-vantage>=2.3.0

# Backtesting Frameworks
backtrader>=1.9.76
vectorbt>=0.25.0

# Technical Analysis
TA-Lib>=0.4.25
ta>=0.10.0

# Optimization
optuna>=3.0.0
scikit-optimize>=0.9.0

# Visualization
matplotlib>=3.5.0
plotly>=5.10.0
seaborn>=0.11.0
mplfinance>=0.12.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.15.0
notebook>=6.4.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.64.0
joblib>=1.1.0

# Web Framework for Mobile App
Flask>=2.3.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Alert System Dependencies
tabulate>=0.9.0

# SMS Alerts
twilio>=8.2.0
schedule>=1.2.0
requests>=2.28.0